+++ DataAddons/Retail/08_BattleForAzeroth/CategoryData.lua
@@ -617,6 +617,8 @@ tinsert(KrowiAF.CategoryData.Expansions, { -- Battle for Azeroth
                 12504, -- The Temple of Sethraliss
                 12505, -- Heroic: The Temple of Sethraliss
                 12506, -- Mythic: The Temple of Sethraliss
+                62443, -- Keystone Hero: Temple of Sethraliss
+                63628, -- Keystone Victor: Temple of Sethraliss
             },
         },
         { -- Tol Dagor
@@ -657,6 +659,8 @@ tinsert(KrowiAF.CategoryData.Expansions, { -- Battle for Azeroth
                 12721, -- Wrap God
                 12723, -- How to Keep a Mummy
                 12848, -- Kings' Rest
+                62444, -- Keystone Hero: Kings' Rest
+                63626, -- Keystone Victor: Kings' Rest
             },
         },
         { -- Operation: Mechagon